Split `WhatIfIndexEstimator` into per-concern modules
