+++ src/Storages/MergeTree/WhatIfStatisticalEstimator.cpp
@@ -1,0 +1,82 @@
+#include <Storages/MergeTree/WhatIfStatisticalEstimator.h>
+
+#include <Storages/MergeTree/WhatIfFilterAnalysis.h>
+#include <Storages/Statistics/ConditionSelectivityEstimator.h>
+#include <Common/Exception.h>
+
+namespace DB
+{
+
+bool tryEstimateWithStatistics(
+    WhatIfIndexEstimator::IndexResult & result,
+    const MergeTreeIndexPtr & index_helper,
+    ReadFromMergeTree * read_step,
+    const ReadFromMergeTree::AnalysisResult & analysis,
+    const RangesInDataParts & parts,
+    const ActionsDAG::Node * filter_node,
+    ContextPtr context)
+{
+    auto metadata = read_step->getStorageMetadata();
+
+    if (!metadata->hasStatistics())
+        return false;
+
+    if (parts.empty())
+        return false;
+
+    /// Only when filter touches just the index's columns, else other columns'
+    /// selectivity leaks into the skip ratio
+    NameSet index_columns_set;
+    for (const auto & col : index_helper->getColumnsRequiredForIndexCalc())
+        index_columns_set.insert(col);
+
+    NameSet filter_input_columns;
+    collectFilterInputColumns(filter_node, filter_input_columns);
+
+    for (const auto & col : filter_input_columns)
+        if (!index_columns_set.contains(col))
+            return false;
+
+    ConditionSelectivityEstimatorBuilder builder(context);
+    bool has_any_stats = false;
+
+    for (const auto & part : parts)
+    {
+        try
+        {
+            auto stats = part.data_part->loadStatistics();
+            if (!stats.empty())
+            {
+                builder.markDataPart(part.data_part);
+                for (const auto & [column_name, stat] : stats)
+                    builder.addStatistics(column_name, stat);
+                has_any_stats = true;
+            }
+        }
+        catch (const Exception &) /// Ok — statistical estimation is best-effort
+        {
+            tryLogCurrentException(__PRETTY_FUNCTION__);
+        }
+    }
+
+    if (!has_any_stats)
+        return false;
+
+    auto estimator = builder.getEstimator();
+    if (!estimator)
+        return false;
+
+    auto profile = estimator->estimateRelationProfile(metadata, filter_node);
+    auto unfiltered = estimator->estimateRelationProfile();
+    if (unfiltered.rows == 0)
+        return false;
+
+    /// Row-level selectivity as upper bound for granule-level skip ratio
+    double selectivity = std::min(1.0, static_cast<double>(profile.rows) / static_cast<double>(unfiltered.rows));
+    result.skip_ratio = 1.0 - selectivity;
+    result.estimated_marks = std::max<UInt64>(1, static_cast<UInt64>(static_cast<double>(analysis.selected_marks) * selectivity));
+    result.estimate_source = "statistical";
+    return true;
+}
+
+}
