+++ src/Storages/MergeTree/WhatIfStatisticalEstimator.h
@@ -1,0 +1,23 @@
+#pragma once
+
+#include <Interpreters/ActionsDAG.h>
+#include <Interpreters/Context_fwd.h>
+#include <Processors/QueryPlan/ReadFromMergeTree.h>
+#include <Storages/MergeTree/MergeTreeIndices.h>
+#include <Storages/MergeTree/RangesInDataPart.h>
+#include <Storages/MergeTree/WhatIfIndexEstimator.h>
+
+namespace DB
+{
+
+/// Estimate skip ratio from column statistics (row-level selectivity as upper bound)
+bool tryEstimateWithStatistics(
+    WhatIfIndexEstimator::IndexResult & result,
+    const MergeTreeIndexPtr & index_helper,
+    ReadFromMergeTree * read_step,
+    const ReadFromMergeTree::AnalysisResult & analysis,
+    const RangesInDataParts & parts,
+    const ActionsDAG::Node * filter_node,
+    ContextPtr context);
+
+}
